Move wee list overview to week folder
